+++ HartreeFock/Core_Start.cpp
@@ -27,7 +27,7 @@ void Core::BuildFirstApproximation(std::string configuration)
     std::string closed_shell_string;
     std::string open_shell_string;
 
-    unsigned int colon_pos = configuration.find(':');
+    size_t colon_pos = configuration.find(':');
     if(colon_pos == std::string::npos)
         closed_shell_string = configuration;
     else
